requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blis==0.7.11
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.16.0
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
dnspython==2.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
fastapi==0.109.2
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
Jinja2==3.1.5
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
MarkupSafe==3.0.2
matplotlib==3.10.1
motor==3.7.0
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.4.2
numpy==1.26.4
packaging==24.2
pandas==2.2.0
pillow==11.1.0
plotly==5.18.0
preshed==3.0.9
PyAudio==0.2.14
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pymongo==4.11.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.10
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
seaborn==0.13.2
setuptools==75.8.2
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.1
srsly==2.5.1
starlette==0.36.3
sympy==1.13.1
telegram==0.0.1
tenacity==9.0.0
thinc==8.2.5
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typer==0.9.4
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wasabi==1.1.3
weasel==0.3.4
